Split console buffer form char device buffer

+++ firmware/app/include/iceCOM.h
@@ -45,10 +45,13 @@ private:
 	std::atomic<bool> m_killThread;
 
 	/* Buffers for char Device transfer */
-	std::vector<char> charDeviceeRx(CHAR_DEVICE_BUFFER);
-	std::vector<char> charDeviceeTx(CHAR_DEVICE_BUFFER);
+	std::vector<char> charDeviceeRx(static_cast<size_t>(CHAR_DEVICE_BUFFER));
+	std::vector<char> charDeviceeTx(static_cast<size_t>(CHAR_DEVICE_BUFFER));
+	// std::vector<char> charDeviceeRx(CHAR_DEVICE_BUFFER);
+	// std::vector<char> charDeviceeTx(CHAR_DEVICE_BUFFER);
 	/* Buffer for the console character input */
-	std::vector<char> consoleBuffer(CONSOLE_BUFFER);
+	std::vector<char> consoleBuffer(static_cast<size_t>(CONSOLE_BUFFER));
+	// std::vector<char> consoleBuffer(CONSOLE_BUFFER);
 
 public:
